Add files via upload
The diet problem is one of the first large-scale optimization problems to be studied in practice. Back in the 1930’s and 40’s, the Army wanted to meet the nutritional requirements of its soldiers while minimizing the cost. This file formulates an optimization model (a linear program) to find the cheapest diet that satisfies the maximum and minimum daily nutrition constraints and solves it using PuLP. Then we introduce the following constraints and solve the new model:
a.	If a food is selected, then a minimum of 1/10 serving must be chosen.
b.	Many people dislike celery and frozen broccoli. So, at most one, but not both, can be selected.
c.	To get day-to-day variety in protein, at least 3 kinds of meat/poultry/fish/eggs must be selected. 

Afterward we solve for the file diet_large.xls, which is a low-cholesterol diet model (rather than minimizing cost, the goal is to minimize cholesterol intake).